Move activity logging into shared module
